requirements.txt
tensorflow==1.15.5
keras==2.3.1
scipy==1.1.0
scikit-learn~=0.24.2
opencv-python~=4.5.3.56
h5py~=2.10.0
matplotlib
Pillow~=7.2.0
requests~=2.26.0
psutil

numpy~=1.21.3
six~=1.15.0
imutils~=0.5.3
torch~=1.9.0
Flask~=2.0.1
sklearn~=0.0
imageio~=2.9.0
pandas~=1.3.1
tqdm~=4.62.0